sanjeevani-backend/app/requirements.txt
# Web Framework
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7

# Database
motor>=3.1.1
pymongo>=4.3.3

# Authentication
firebase-admin>=6.1.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6

# Integrations
googlemaps>=4.10.0
twilio>=8.0.0
httpx>=0.24.0

# Utilities
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# Multilingual Support
googletrans>=3.1.0a0
deep-translator>=1.11.1

# Testing
pytest>=7.3.1
httpx>=0.24.0

# CORS
fastapi-cors>=0.0.6